BloggingAPI/requirements.txt
# Install with: `pip install -r requirements.txt`

django==1.9
djangorestframework
psycopg2
pillow
drf-nested-routers
django-cors-headers
django-rest-auth
django-allauth
